requirements.txt
tensorflow==2.9.1
torch==1.12.1
torchvision==0.13.1
Flask==2.2.2
opencv-python==4.5.5.64
Pillow==9.2.0
easyocr==1.5.0
gunicorn==19.9.0